mudando para integer em todas as classes

+++ src/main/java/br/com/fiap/apisecurity/dto/usuario/LoginResponse.java
@@ -3,10 +3,10 @@ package br.com.fiap.apisecurity.dto.usuario;
 public class LoginResponse {
 
     private String token;
-    private Long idUsuario;
-    private Long idTipoUsuario;
+    private Integer idUsuario;
+    private Integer idTipoUsuario;
 
-    public LoginResponse(String token, Long idUsuario, Long idTipoUsuario) {
+    public LoginResponse(String token, Integer idUsuario, Integer idTipoUsuario) {
         this.token = token;
         this.idUsuario = idUsuario;
         this.idTipoUsuario = idTipoUsuario;
@@ -16,11 +16,11 @@ public class LoginResponse {
         return token;
     }
 
-    public Long getIdUsuario() {
+    public Integer getIdUsuario() {
         return idUsuario;
     }
 
-    public Long getIdTipoUsuario() {
+    public Integer getIdTipoUsuario() {
         return idTipoUsuario;
     }
 }
